requirements.txt
pocketflow>=0.0.1
openai>=1.0.0
numpy>=1.19.0
faiss-cpu>=1.7.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0